[FIX] 차단 삭제 API 수정
제대로 된 blockId, 유저가 작성한 block인지 판단하도록

+++ src/main/java/org/websoso/WSSServer/service/BlockService.java
@@ -2,6 +2,7 @@ package org.websoso.WSSServer.service;
 
 import static org.websoso.WSSServer.exception.block.BlockErrorCode.ALREADY_BLOCKED;
 import static org.websoso.WSSServer.exception.block.BlockErrorCode.BLOCK_NOT_FOUND;
+import static org.websoso.WSSServer.exception.block.BlockErrorCode.INVALID_AUTHORIZED_BLOCK;
 import static org.websoso.WSSServer.exception.block.BlockErrorCode.SELF_BLOCKED;
 
 import java.util.List;
@@ -15,6 +16,7 @@ import org.websoso.WSSServer.dto.block.BlockGetResponse;
 import org.websoso.WSSServer.dto.block.BlocksGetResponse;
 import org.websoso.WSSServer.exception.block.exception.AlreadyBlockedException;
 import org.websoso.WSSServer.exception.block.exception.BlockNotFoundException;
+import org.websoso.WSSServer.exception.block.exception.InvalidAuthorizedBlockException;
 import org.websoso.WSSServer.exception.block.exception.SelfBlockedException;
 import org.websoso.WSSServer.repository.BlockRepository;
 
@@ -54,8 +56,12 @@ public class BlockService {
 
     @Transactional
     public void deleteBlock(User user, Long blockId) {
-        Block block = blockRepository.findByBlockingIdAndBlockedId(user.getUserId(), blockId).orElseThrow(() ->
+        Block block = blockRepository.findById(blockId).orElseThrow(() ->
                 new BlockNotFoundException(BLOCK_NOT_FOUND, "block with the given blockId was not found"));
+        if (!block.getBlockingId().equals(user.getUserId())) {
+            throw new InvalidAuthorizedBlockException(INVALID_AUTHORIZED_BLOCK,
+                    "block with the given blockId is not from user with the given userId");
+        }
         blockRepository.delete(block);
     }
 }
